requirements.txt
alembic==1.3.0
boto3==1.10.14
celery==4.3.0
Click==7.0
Flask==1.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
sendgrid==6.1.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1

pendulum~=2.1.2
phonenumbers~=8.12.17
twilio~=6.51.0
stripe~=2.55.1
pgeocode~=0.3.0